feat: dynamic collection schemas via CUE config
Add `fields` to CollectionConfig in CUE schema so users define
per-collection frontmatter field schemas (Typesense types, facet, sort).
Fields are wired into Typesense schema creation (create + additive update)
and frontmatter values are extracted at index time.

- CUE: CollectionConfig.fields, FrontmatterField type float64 -> float
- config: CollectionConfig.Fields map[string]FrontmatterField
- ts: SchemaField type, EnsureSchema create/update, GetSchemaFields,
  ChunkDocument.Fields + ToMap for dynamic field merging
- runtime: buildFieldSchemas merges + validates types across collections
- indexer: parse frontmatter when collection defines fields, store in doc
- wiki: validate/filter use 'float' instead of 'float64'
- collection show: prints configured fields and TS schema diff
- doctor: validates schema sync (PENDING/ORPHAN/TYPEMISMATCH)
- project config: commented example of fields usage
- AGENTS.md: integration test timing guidance

+++ pkg/wiki/frontmatter_integration_test.go
@@ -136,7 +136,7 @@ func TestIntegrationValidateFrontmatter_IntField(t *testing.T) {
 func TestIntegrationValidateFrontmatter_FloatField(t *testing.T) {
 	cfg := &config.FrontmatterConfig{
 		Fields: map[string]config.FrontmatterField{
-			"score": {Type: "float64"},
+			"score": {Type: "float"},
 		},
 	}
 	if err := ValidateFrontmatter(map[string]interface{}{"score": 3.14}, cfg); err != nil {
@@ -218,7 +218,7 @@ func TestIntegrationFrontmatterToFilter_IntField(t *testing.T) {
 func TestIntegrationFrontmatterToFilter_FloatField(t *testing.T) {
 	cfg := &config.FrontmatterConfig{
 		Fields: map[string]config.FrontmatterField{
-			"score": {Type: "float64"},
+			"score": {Type: "float"},
 		},
 	}
 	s := FrontmatterToFilter(map[string]interface{}{"score": 3.14}, cfg)
